Remove settings.py
Why: The file settings.py was containing the image_user pwd.

This file has been removed from all the git history (with bfg) and
and a new "settings.py_SAMPLE" file has been added to git,
with "**REMOVED**" instead of the password.

Developers should copy settings.py_SAMPLE to settings.py everytime a change is
made in the settings.py_SAMPLE file,
and write the password into the settings.py file.

Changes to be committed:
	deleted:    settings.py
	modified:   .gitignore
	new file:   django-data/image/image/settings.py_SAMPLE