Add organization_role and person_role
An organization.role field, with a foreign key to a roles table was
already present. Here, we added a varchar(255) organization_role that will
be the only one used in the near future.

A person.role field, with a foreign key to a roles table was
already present. Here, we added a varchar(255) person_role that will
be the only one used in the near future.

The two new fields have been added to the admin.py file