:bug: bug fixed in validation.tasks.ValidateTask
http://docs.celeryproject.org/en/latest/userguide/tasks.html#instantiation
A task is not instantiated for every request, but is registered in
the task registry as a global instance. This means that the __init__
constructor will only be called once per process, and that the
task class is semantically closer to an Actor. if you have a task and
you route every request to the same process, then it will keep state
between requests. This can also be useful to cache resources, For
example, a base Task class that caches a database connection